--- index.js
@@ -1,7 +1,0 @@
-const { loadImage } = require('canvas');
-
-module.exports = {
-	CaptchaGenerator: require('./src/main'),
-	version: require('./package.json').version,
-	resolveImage: loadImage,
-};
