+++ index.ts
@@ -1,0 +1,5 @@
+import { loadImage } from 'canvas';
+import CaptchaGenerator from './src/main';
+import { version } from './package.json';
+
+export { CaptchaGenerator, version, loadImage as resolveImage };
