+++ package.json
@@ -2,8 +2,9 @@
   "name": "captcha-canvas",
   "version": "2.4.0",
   "description": "A captcha generator by using canvas module.",
-  "main": "index.js",
+  "main": "dist/index.js",
   "scripts": {
+    "build": "tsc",
     "docsgen": "jsdoc --configure .jsdoc.json --verbose",
     "example": "node examples/example.js",
     "start": "node index.js",
@@ -31,15 +32,15 @@
     "canvas": "^2.7.0"
   },
   "devDependencies": {
-    "@types/node": "^14.11.8",
+    "@types/node": "^18.19.120",
     "eslint": "^7.27.0",
     "jsdoc": "^3.6.5",
     "jsdoc-skyceil": "^1.0.5",
-    "typescript": "^4.0.2"
+    "typescript": "^4.9.4"
   },
   "directories": {
     "doc": "docs",
     "example": "examples"
   },
-  "types": "./typings/index.d.ts"
+  "types": "./dist/index.d.ts"
 }
